New module Alphabetable with alphabet methods from Encryption moved to it
