+++ lib/alphabetable.rb
@@ -1,0 +1,14 @@
+module Alphabetable
+  def alphabet
+    ("a".."z").to_a << " "  
+  end
+
+  def alphabet_with_index
+    index_hash = Hash.new(0)
+    alphabet.each_with_index do |letter, index|
+      index_hash["#{letter}"] = "#{index}".to_i
+    end
+    index_hash
+  end
+
+end
